Ignore past events returned by the Meetup API
Summary:
The meetup API `/events` endpoint is returning completed events with the
status "upcoming". An [issue][1] has been submitted to the meetup API
repo but in the interim, we want to prevent the upcoming event section
from showing an event that is from a previous day.

In addition to filtering out past events, this PR takes the opportunity
to refactor the existing code by:
- introducing an EventParser
- introducing an EventPresenter
- renaming MeetupEvent to Event
- adding the Gem attr_extras

[1]: https://github.com/meetup/api/issues/301